adding hooks for extras and other modules
