+++ my-video-room/dao/class-roomadmin.php
@@ -14,7 +14,6 @@ use MyVideoRoomPlugin\Library\Dependencies;
 use MyVideoRoomPlugin\Shortcode as Shortcode;
 use MyVideoRoomPlugin\Factory;
 use MyVideoRoomPlugin\DAO\ModuleConfig;
-use MyVideoRoomPlugin\Module\SiteVideo\MVRSiteVideo;
 
 /**
  * Class RoomAdmin
@@ -58,7 +57,7 @@ class RoomAdmin extends Shortcode {
 				return $post_id;
 			case 'type':
 					$room = Factory::get_instance( RoomMap::class )->get_room_info( $room_post_id );
-				return $this->conference_room_friendly_name( $room->room_type );
+				return \apply_filters( 'mvr_room_type_display_override', $room->room_type );
 			case 'post_id':
 				return $post_id;
 			case 'title':
@@ -143,24 +142,5 @@ class RoomAdmin extends Shortcode {
 		);
 
 		return (bool) $current_user_setting;
-	}
-
-	/**
-	 * Room Type Friendly Name
-	 *
-	 * @param string $room_type .
-	 * @return string name.
-	 */
-	public function conference_room_friendly_name( string $room_type ): string {
-		switch ( $room_type ) {
-			case MVRSiteVideo::ROOM_SHORTCODE_SITE_VIDEO:
-				if ( ! Factory::get_instance( ModuleConfig::class )->read_enabled_status( MVRSiteVideo::MODULE_SITE_VIDEO_ID ) ) {
-					return esc_html__( 'Module Disabled', 'myvideoroom' );
-				} else {
-					return MVRSiteVideo::ROOM_NAME_TABLE;
-				}
-		}
-
-		return $room_type;
 	}
 }
