+++ my-video-room/dao/class-roommap.php
@@ -228,15 +228,22 @@ class RoomMap {
 	 *
 	 * @return array
 	 */
-	public function get_room_list( string $room_type ): array {
+	public function get_room_list( string $room_type = null ): array {
 		global $wpdb;
-
-		$raw_sql = '
+		if ( ! $room_type ){
+			$raw_sql = '
+			SELECT post_id
+			FROM ' . $wpdb->prefix . self::TABLE_NAME . '
+			ORDER BY room_type ASC
+		';
+		} else {
+			$raw_sql = '
 			SELECT post_id
 			FROM ' . $wpdb->prefix . self::TABLE_NAME . '
 			WHERE room_type = %s
+			ORDER BY room_type ASC
 		';
-
+		}
 		$prepared_query = $wpdb->prepare(
 			// phpcs:ignore WordPress.DB.PreparedSQL.NotPrepared
 			$raw_sql,
