+++ my-video-room/module/sitevideo/library/class-mvrsitevideoviews.php
@@ -7,6 +7,7 @@
 
 namespace MyVideoRoomPlugin\Module\SiteVideo\Library;
 
+use MyVideoRoomPlugin\DAO\ModuleConfig;
 use MyVideoRoomPlugin\Factory;
 use MyVideoRoomPlugin\Shortcode as Shortcode;
 use MyVideoRoomPlugin\DAO\RoomMap;
@@ -72,5 +73,21 @@ class MVRSiteVideoViews extends Shortcode {
 		return $render( $module_id, $name_output, $host_id, $room_name, $is_guest, $meeting_link, $module_suffix );
 
 	}
-
+	/**
+	 * Room Type Friendly Name
+	 *
+	 * @param string $room_type .
+	 * @return string name.
+	 */
+	public function conference_room_friendly_name( string $room_type ): string {
+		switch ( $room_type ) {
+			case MVRSiteVideo::ROOM_SHORTCODE_SITE_VIDEO:
+				if ( ! Factory::get_instance( ModuleConfig::class )->read_enabled_status( MVRSiteVideo::MODULE_SITE_VIDEO_ID ) ) {
+					return esc_html__( 'Module Disabled', 'myvideoroom' );
+				} else {
+					return MVRSiteVideo::ROOM_NAME_TABLE;
+				}
+		}
+		return $room_type;
+	}
 }
